Initial stab at a Ruby gem with embedded Rust component.
This wraps my Rust implementation of `.gitignore` semantics with a small
C ABI compatable interface which gets compiled as a gem with extensions.

Unfortunately, it currently requires nightly Rust as it uses parts of
Rust `libc` marked as unstable, but hopefully we can get around this at
some point too.